Add HathiTrust cover image fetcher
This retrieves the probable best cover from the HathiTrust Data API, retrieving the volume's structure in a METS document and pulling out the best page. This adds some time to the processing of the daily files so the number of threads has been bumped up to handle the increased processing time. It should not cause any timeout issues, but if it does those should be addressable by increasing the memory of the AWS instance (this runs once a day so cost shouldn't be a concern, even at a max state the monthly cost is less than $5)

The URLs retrieved are temporarily stored in the database before being replaced with permanent URLs that point to local files fetched as part of the follow on ingest process.