Recursively handle batches of Hathi Records
The new data enhancement steps (mainly the fetching of VIAF numbers at this phase) have caused a large number of timeouts in this function as it cannot complete its execution in the 15 minute window for lambda functions.

This splits the received records into chunks of 500 and then recursively calls itself to process the remaining records. This limits the individual execution itme for each call and increases the parallelization so that the overall execution time should take no longer than 15 minutes